+++ camerax/example/lib/models.dart
@@ -1,3 +1,2 @@
 export 'models/camera_mode.dart';
 export 'models/image_model.dart';
-export 'models/limited_value.dart';