+++ camerax/example/lib/views/faces_view.dart
@@ -37,7 +37,6 @@ class _FacesPainter extends CustomPainter {
 
   @override
   void paint(Canvas canvas, ui.Size size) {
-    debugPrint('DRAW ON $size');
     final paint = Paint()
       ..style = PaintingStyle.stroke
       ..strokeJoin = StrokeJoin.bevel