+++ camerax_ios/ios/camerax_ios/Sources/camerax_ios/avfoundation/AVMetadataObjectDelegate.swift
@@ -10,10 +10,10 @@ import AVFoundation
 
 class AVMetadataObjectDelegate: PigeonApiDelegateAVMetadataObjectProxyApi {
     func getType(pigeonApi: PigeonApiAVMetadataObjectProxyApi, pigeonInstance: AVMetadataObject) throws -> AVMetadataObjectTypeApi {
-        guard let api = pigeonInstance.type.api else {
+        guard let type = pigeonInstance.type.apiOrNil else {
             throw CameraXError(code: "nil-error", message: "type is nil", details: nil)
         }
-        return api
+        return type
     }
     
     func getTime(pigeonApi: PigeonApiAVMetadataObjectProxyApi, pigeonInstance: AVMetadataObject) throws -> Int64 {
@@ -38,7 +38,7 @@ class AVMetadataObjectDelegate: PigeonApiDelegateAVMetadataObjectProxyApi {
     
     func getCinematicVideoFocusMode(pigeonApi: PigeonApiAVMetadataObjectProxyApi, pigeonInstance: AVMetadataObject) throws -> AVCaptureDeviceCinematicVideoFocusModeApi {
         if #available(iOS 26.0, *) {
-            guard let mode = pigeonInstance.cinematicVideoFocusMode.api else {
+            guard let mode = pigeonInstance.cinematicVideoFocusMode.apiOrNil else {
                 throw CameraXError(code: "nil-error", message: "cinematicVideoFocusMode is nil", details: nil)
             }
             return mode
@@ -67,7 +67,7 @@ class AVMetadataObjectDelegate: PigeonApiDelegateAVMetadataObjectProxyApi {
 }
 
 extension AVMetadataObject.ObjectType {
-    var api: AVMetadataObjectTypeApi? {
+    var apiOrNil: AVMetadataObjectTypeApi? {
         if #available(iOS 26.0, *) {
             switch self {
             case .dogHead: return .dogHead
@@ -117,51 +117,35 @@ extension AVMetadataObject.ObjectType {
 }
 
 extension AVMetadataObjectTypeApi {
-    var impl: AVMetadataObject.ObjectType? {
-        if #available(iOS 26.0, *) {
-            switch self {
-            case .dogHead: return .dogHead
-            case .catHead: return .catHead
-            default: break
-            }
-        }
-        if #available(iOS 17.0, *) {
-            switch self {
-            case .humanFullBody: return .humanFullBody
-            default: break
-            }
-        }
-        if #available(iOS 15.4, *) {
-            switch self {
-            case .codabar: return .codabar
-            case .gs1DataBar: return .gs1DataBar
-            case .gs1DataBarExpanded: return .gs1DataBarExpanded
-            case .gs1DataBarLimited: return .gs1DataBarLimited
-            case .microPDF417: return .microPDF417
-            case .microQR: return .microQR
-            default: break
-            }
-        }
+    var delegateOrNil: AVMetadataObject.ObjectType? {
         return switch self {
-        case .humanBody: .humanBody
-        case .dogBody: .dogBody
-        case .catBody: .catBody
-        case .salientObject: .salientObject
+        case .codabar: if #available(iOS 15.4, *) { .codabar } else { nil }
         case .code39: .code39
         case .code39Mod43: .code39Mod43
         case .code93: .code93
         case .code128: .code128
         case .ean8: .ean8
         case .ean13: .ean13
+        case .gs1DataBar: if #available(iOS 15.4, *) { .gs1DataBar } else { nil }
+        case .gs1DataBarExpanded: if #available(iOS 15.4, *) { .gs1DataBarExpanded } else { nil }
+        case .gs1DataBarLimited: if #available(iOS 15.4, *) { .gs1DataBarLimited } else { nil }
         case .interleaved2of5: .interleaved2of5
         case .itf14: .itf14
         case .upce: .upce
         case .aztec: .aztec
         case .dataMatrix: .dataMatrix
+        case .microPDF417: if #available(iOS 15.4, *) { .microPDF417 } else { nil }
+        case .microQR: if #available(iOS 15.4, *) { .microQR } else { nil }
         case .pdf417: .pdf417
         case .qr: .qr
+        case .humanBody: .humanBody
+        case .humanFullBody: if #available(iOS 17.0, *) { .humanFullBody } else { nil }
+        case .dogHead: if #available(iOS 26.0, *) { .dogHead } else { nil }
+        case .dogBody: .dogBody
+        case .catHead: if #available(iOS 26.0, *) { .catHead } else { nil }
+        case .catBody: .catBody
         case .face: .face
-        default: nil
+        case .salientObject: .salientObject
         }
     }
 }