+++ camerax_ios/ios/camerax_ios/Sources/camerax_ios/common/ImageFormatDelegate.swift
@@ -9,93 +9,50 @@ import Foundation
 
 extension ImageFormat {
     var api: ImageFormatApi {
-        switch self {
-        case .yuv420_888:
-            return .yuv420888
-        case .rgba8888:
-            return .rgba8888
-        case .jpeg:
-            return .jpeg
-        case .depth16:
-            return .depth16
-        case .depthJpeg:
-            return .depthJpeg
-        case .depthPointCloud:
-            return .depthPointCloud
-        case .flexRgba8888:
-            return .flexRgba8888
-        case .flexRgb888:
-            return .flexRgb888
-        case .heic:
-            return .heic
-        case .heicUltraHdr:
-            return .heicUltraHdr
-        case .jpegR:
-            return .jpegR
-        case .nv16:
-            return .nv16
-        case .nv21:
-            return .nv21
-        case .androidPrivate:
-            return .androidPrivate
-        case .raw10:
-            return .raw10
-        case .raw12:
-            return .raw12
-        case .rawPrivate:
-            return .rawPrivate
-        case .rawSensor:
-            return .rawSensor
-        case .unknown:
-            return .unknown
-        case .y8:
-            return .y8
-        case .yCbCrP010:
-            return .yCbCrP010
-        case .yCbCrP210:
-            return .yCbCrP210
-        case .yuv422_888:
-            return .yuv422888
-        case .yuv444_888:
-            return .yuv444888
-        case .yuy2:
-            return .yuy2
-        case .yv12:
-            return .yv12
-        case .a8:
-            return .a8
-        case .la88:
-            return .la88
-        case .l8:
-            return .l8
-        case .opaque:
-            return .opaque
-        case .rgba1010102:
-            return .rgba1010102
-        case .rgba4444:
-            return .rgba4444
-        case .rgba5551:
-            return .rgba5551
-        case .rgbaF16:
-            return .rgbaF16
-        case .rgbx8888:
-            return .rgbx8888
-        case .rgb332:
-            return .rgb332
-        case .rgb565:
-            return .rgb565
-        case .rgb888:
-            return .rgb888
-        case .translucent:
-            return .translucent
-        case .transparent:
-            return .transparent
-        case .yCbCr420Sp:
-            return .yCbCr420Sp
-        case .yCbCr422I:
-            return .yCbCr422I
-        case .yCbCr422Sp:
-            return .yCbCr422Sp
+        return switch self {
+        case .yuv420_888: .yuv420888
+        case .rgba8888: .rgba8888
+        case .jpeg: .jpeg
+        case .depth16: .depth16
+        case .depthJpeg: .depthJpeg
+        case .depthPointCloud: .depthPointCloud
+        case .flexRgba8888: .flexRgba8888
+        case .flexRgb888: .flexRgb888
+        case .heic: .heic
+        case .heicUltraHdr: .heicUltraHdr
+        case .jpegR: .jpegR
+        case .nv16: .nv16
+        case .nv21: .nv21
+        case .androidPrivate: .androidPrivate
+        case .raw10: .raw10
+        case .raw12: .raw12
+        case .rawPrivate: .rawPrivate
+        case .rawSensor: .rawSensor
+        case .unknown: .unknown
+        case .y8: .y8
+        case .yCbCrP010: .yCbCrP010
+        case .yCbCrP210: .yCbCrP210
+        case .yuv422_888: .yuv422888
+        case .yuv444_888: .yuv444888
+        case .yuy2: .yuy2
+        case .yv12: .yv12
+        case .a8: .a8
+        case .la88: .la88
+        case .l8: .l8
+        case .opaque: .opaque
+        case .rgba1010102: .rgba1010102
+        case .rgba4444: .rgba4444
+        case .rgba5551: .rgba5551
+        case .rgbaF16: .rgbaF16
+        case .rgbx8888: .rgbx8888
+        case .rgb332: .rgb332
+        case .rgb565: .rgb565
+        case .rgb888: .rgb888
+        case .translucent: .translucent
+        case .transparent: .transparent
+        case .yCbCr420Sp: .yCbCr420Sp
+        case .yCbCr422I: .yCbCr422I
+        case .yCbCr422Sp: .yCbCr422Sp
         }
     }
 }