+++ camerax_ios/lib/src/common/view_impl.dart
@@ -1,5 +1,4 @@
 import 'package:camerax_platform_interface/camerax_platform_interface.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart' hide View;
 
@@ -9,48 +8,11 @@ base mixin ViewImpl on View {
 
   @override
   Widget build(BuildContext context) {
-    return PlatformViewLink(
+    return UiKitView(
       viewType: viewType,
-      surfaceFactory: (context, controller) => AndroidViewSurface(
-        controller: controller as AndroidViewController,
-        hitTestBehavior: PlatformViewHitTestBehavior.opaque,
-        gestureRecognizers: const {},
-      ),
-      onCreatePlatformView: (params) =>
-          initAndroidView(
-              params,
-              hybridComposition: false,
-              creationParams: identifier,
-              creationParamsCodec: const StandardMessageCodec(),
-            )
-            ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
-            ..create(),
+      layoutDirection: TextDirection.ltr,
+      creationParams: identifier,
+      creationParamsCodec: const StandardMessageCodec(),
     );
-  }
-
-  AndroidViewController initAndroidView(
-    PlatformViewCreationParams params, {
-    required bool hybridComposition,
-    TextDirection layoutDirection = TextDirection.ltr,
-    dynamic creationParams,
-    MessageCodec<dynamic>? creationParamsCodec,
-  }) {
-    if (hybridComposition) {
-      return PlatformViewsService.initExpensiveAndroidView(
-        id: params.id,
-        viewType: params.viewType,
-        layoutDirection: layoutDirection,
-        creationParams: creationParams,
-        creationParamsCodec: creationParamsCodec,
-      );
-    } else {
-      return PlatformViewsService.initSurfaceAndroidView(
-        id: params.id,
-        viewType: params.viewType,
-        layoutDirection: layoutDirection,
-        creationParams: creationParams,
-        creationParamsCodec: creationParamsCodec,
-      );
-    }
   }
 }